Web Inspector: CodeGeneratorInspector.py: protect typed API from C++ implicit float to int cast https://bugs.webkit.org/show_bug.cgi?id=87183
Patch by Peter Rybin <peter.rybin@gmail.com> on 2012-05-28
Reviewed by Yury Semikhatsky.

An intermediate C++ class is introduced that uses C++ template technique to control actual type
of its constructor argument.
All input parameters of type "int" now have type ExactlyInt.
All usage sites are fixed accordingly.

* inspector/CodeGeneratorInspector.py:
(TypeModel.RefPtrBased):
(TypeModel.Enum):
(TypeModel.ValueType):
(TypeModel.ValueType.get_opt_output_type_):
(TypeModel.ValueType.ValueOptional.get_command_return_pass_model):
(TypeModel.ExactlyInt):
(TypeModel.ExactlyInt.__init__):
(TypeModel.ExactlyInt.get_input_param_type_text):
(TypeModel.ExactlyInt.get_opt_output_type_):
(TypeModel.init_class):
(ExactlyInt):
* inspector/InspectorAgent.cpp:
(WebCore::InspectorAgent::enable):
(WebCore::InspectorAgent::didCreateWorker):
(WebCore::InspectorAgent::didDestroyWorker):
(WebCore::InspectorAgent::evaluateForTestInFrontend):
* inspector/InspectorApplicationCacheAgent.cpp:
(WebCore::InspectorApplicationCacheAgent::updateApplicationCacheStatus):
(WebCore::InspectorApplicationCacheAgent::getFramesWithManifests):
* inspector/InspectorDOMAgent.cpp:
(WebCore::InspectorDOMAgent::buildObjectForNode):
* inspector/InspectorMemoryAgent.cpp:
(WebCore::jsHeapInfo):
(WebCore::InspectorMemoryAgent::getProcessMemoryDistribution):
* inspector/PageRuntimeAgent.cpp:
(WebCore::PageRuntimeAgent::notifyContextCreated):

Canonical link: https://commits.webkit.org/105397@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118652 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/inspector/InspectorAgent.cpp
@@ -126,7 +126,7 @@ void InspectorAgent::enable(ErrorString*)
     WorkersMap::iterator workersEnd = m_workers.end();
     for (WorkersMap::iterator it = m_workers.begin(); it != workersEnd; ++it) {
         InspectorWorkerResource* worker = it->second.get();
-        m_frontend->inspector()->didCreateWorker(worker->id(), worker->url(), worker->isSharedWorker());
+        m_frontend->inspector()->didCreateWorker(static_cast<int>(worker->id()), worker->url(), worker->isSharedWorker());
     }
 #endif
 
@@ -134,7 +134,7 @@ void InspectorAgent::enable(ErrorString*)
         inspect(m_pendingInspectData.first, m_pendingInspectData.second);
 
     for (Vector<pair<long, String> >::iterator it = m_pendingEvaluateTestCommands.begin(); m_frontend && it != m_pendingEvaluateTestCommands.end(); ++it)
-        m_frontend->inspector()->evaluateForTestInFrontend((*it).first, (*it).second);
+        m_frontend->inspector()->evaluateForTestInFrontend(static_cast<int>((*it).first), (*it).second);
     m_pendingEvaluateTestCommands.clear();
 }
 
@@ -163,7 +163,7 @@ void InspectorAgent::didCreateWorker(intptr_t id, const String& url, bool isShar
     m_workers.set(id, workerResource);
 #if ENABLE(JAVASCRIPT_DEBUGGER)
     if (m_inspectedPage && m_frontend && m_state->getBoolean(InspectorAgentState::inspectorAgentEnabled))
-        m_frontend->inspector()->didCreateWorker(id, url, isSharedWorker);
+        m_frontend->inspector()->didCreateWorker(static_cast<int>(id), url, isSharedWorker);
 #endif
 }
 
@@ -177,7 +177,7 @@ void InspectorAgent::didDestroyWorker(intptr_t id)
         return;
 #if ENABLE(JAVASCRIPT_DEBUGGER)
     if (m_inspectedPage && m_frontend && m_state->getBoolean(InspectorAgentState::inspectorAgentEnabled))
-        m_frontend->inspector()->didDestroyWorker(id);
+        m_frontend->inspector()->didDestroyWorker(static_cast<int>(id));
 #endif
     m_workers.remove(workerResource);
 }
@@ -186,7 +186,7 @@ void InspectorAgent::didDestroyWorker(intptr_t id)
 void InspectorAgent::evaluateForTestInFrontend(long callId, const String& script)
 {
     if (m_state->getBoolean(InspectorAgentState::inspectorAgentEnabled))
-        m_frontend->inspector()->evaluateForTestInFrontend(callId, script);
+        m_frontend->inspector()->evaluateForTestInFrontend(static_cast<int>(callId), script);
     else
         m_pendingEvaluateTestCommands.append(pair<long, String>(callId, script));
 }
